Move test to more appropriate module
